# backend/app/schemas/order.py
from datetime import datetime
from decimal import Decimal
from uuid import UUID

from pydantic import BaseModel, ConfigDict, Field

from app.core.schema import TimestampSchemaRead
from app.models.order import OrderStatus, PaymentStatus
from app.schemas.user import PublicUserRead

# ─── OrderItem Schemas ──────────────────────────────────────────────────────


class OrderItemBase(BaseModel):
    product_id: int | None = None
    variant_id: int | None = None
    course_id: int | None = None
    quantity: int = 1
    price: Decimal


class OrderItemCreate(OrderItemBase):
    order_id: UUID


class OrderItemRead(OrderItemBase):
    id: int
    order_id: UUID
    model_config = ConfigDict(from_attributes=True)


# ─── Order Schemas ──────────────────────────────────────────────────────────


class OrderBase(BaseModel):
    shipping_address_id: int | None = None
    billing_address_id: int | None = None
    payment_method: str = "razorpay"


class OrderCreateDB(OrderBase):
    user_id: UUID
    total_amount: Decimal


class OrderCreate(OrderBase):
    items: list[OrderItemBase]


class OrderUpdate(BaseModel):
    status: OrderStatus | None = None
    payment_status: PaymentStatus | None = None
    razorpay_payment_id: str | None = None
    razorpay_signature: str | None = None


class OrderRead(OrderBase, TimestampSchemaRead):
    id: UUID
    user_id: UUID
    status: OrderStatus
    payment_status: PaymentStatus
    razorpay_order_id: str | None = None
    total_amount: Decimal
    items: list[OrderItemRead] = []

    model_config = ConfigDict(from_attributes=True)


class AdminOrderRead(OrderRead):
    user: PublicUserRead | None = None


class OrderDashboardRead(TimestampSchemaRead):
    id: UUID
    user: PublicUserRead | None = None
    total_amount: Decimal
    status: OrderStatus
    payment_status: PaymentStatus
    item_count: int

    model_config = ConfigDict(from_attributes=True)


class OrderListParams(BaseModel):
    page: int = Field(1, ge=1)
    page_size: int = Field(10, ge=1)
    search: str | None = None
    status: OrderStatus | None = None
    payment_status: PaymentStatus | None = None
    date_start: datetime | None = None
    date_end: datetime | None = None
    min_amount: Decimal | None = Field(None, ge=0)
    max_amount: Decimal | None = Field(None, ge=0)

    model_config = {"extra": "forbid"}


# ─── Razorpay Verification ──────────────────────────────────────────────────


class PaymentVerificationRequest(BaseModel):
    razorpay_order_id: str
    razorpay_payment_id: str
    razorpay_signature: str